added necessary math include

+++ src/examples/markdw1000driver/DW1000Device.cpp
@@ -23,6 +23,7 @@
 
 #include "DW1000Device.h"
 #include "DW1000.h"
+#include <math.h>
 
 
 //Constructor and destructor
@@ -66,13 +67,13 @@ void DW1000Device::setShortAddress(uint8_t deviceAddress[]) {
 }
 
 
-void DW1000Device::setRange(float range) { _range = round(range*100); }
+void DW1000Device::setRange(float range) { _range = roundf(range*100); }
 
-void DW1000Device::setRXPower(float RXPower) { _RXPower = round(RXPower*100); }
+void DW1000Device::setRXPower(float RXPower) { _RXPower = roundf(RXPower*100); }
 
-void DW1000Device::setFPPower(float FPPower) { _FPPower = round(FPPower*100); }
+void DW1000Device::setFPPower(float FPPower) { _FPPower = roundf(FPPower*100); }
 
-void DW1000Device::setQuality(float quality) { _quality = round(quality*100); }
+void DW1000Device::setQuality(float quality) { _quality = roundf(quality*100); }
 
 
 uint8_t* DW1000Device::getByteAddress() {
